Debug mode working. Made a real progress bar.
Still need to test if the dev data makes it in the public version of the
website (aka when debug mode is not on). Otherwise this should speed
development up.

Started a components js folder. Still need to read about the proper way
to create them tho, so the structure might change.